requirements.txt
Django==4.2.0
psycopg2-binary
django-environ
django-polymorphic
